+++ components/applied-filters/applied-filters.js
