+++ components/search-facets/search-facets-localize-behavior.js
