+++ components/search-facets/search-facets-option.js
