+++ components/search-results-count/search-results-count-localize-behavior.js
