+++ components/sort-by-dropdown/sort-by-dropdown-localize-behavior.js
