+++ demo/applied-filters/applied-filters.html
@@ -1,0 +1,90 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script type="module">
+			import '@brightspace-ui/core/components/demo/demo-page.js';
+			import '@brightspace-ui/core/components/typography/typography.js';
+			import '../../components/filter-dropdown/filter-dropdown.js';
+			import '../../components/filter-dropdown/filter-dropdown-category.js';
+			import '../../components/filter-dropdown/filter-dropdown-option.js';
+			import '../../components/applied-filters/applied-filters.js';
+		</script>
+		<style>
+			d2l-filter-dropdown, d2l-applied-filters { display: block; }
+		</style>
+	</head>
+	<body unresolved class="d2l-typography">
+
+		<d2l-demo-page page-title="d2l-applied-filters">
+
+			<h2>Use this dropdown filter to control the applied filters:</h2>
+
+			<d2l-filter-dropdown id="filter" total-selected-option-count="2">
+				<d2l-filter-dropdown-category key="1" category-text="Cat1">
+					<d2l-filter-dropdown-option selected text="Option 1 - 1" value="1"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 1 - 4" value="4"></d2l-filter-dropdown-option>
+				</d2l-filter-dropdown-category>
+				<d2l-filter-dropdown-category key="2" category-text="Cat2">
+					<d2l-filter-dropdown-option selected text="Option 2 - 1" value="1"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 2 - 2" value="2"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 2 - 3" value="3"></d2l-filter-dropdown-option>
+				</d2l-filter-dropdown-category>
+			</d2l-filter-dropdown>
+			<script>
+				const filter = document.querySelector('#filter');
+				filter.addEventListener('d2l-filter-dropdown-close', (e) => {
+					logEvent(e, 'Filter dropdown closed!');
+				});
+				filter.addEventListener('d2l-filter-dropdown-cleared', (e) => {
+					logEvent(e, 'Filters clear button clicked!');
+				});
+				filter.addEventListener('d2l-filter-dropdown-category-selected', (e) => {
+					logEvent(e, 'Filter category tab selected!');
+				});
+				filter.addEventListener('d2l-filter-dropdown-category-searched', (e) => {
+					logEvent(e, 'Filter category searched!');
+				});
+				filter.addEventListener('d2l-filter-dropdown-category-slotchange', (e) => {
+					logEvent(e, 'Filter category slotchange!');
+				});
+				filter.addEventListener('d2l-filter-dropdown-option-change', (e) => {
+					logEvent(e, 'Filter option clicked!');
+					if (e.detail.selected) {
+						filter.totalSelectedOptionCount++;
+					} else {
+						filter.totalSelectedOptionCount--;
+					}
+				});
+				/* eslint-disable no-console */
+				function logEvent(e, text) {
+					console.group(text);
+					console.log('event', e);
+					if (e.detail) console.log('detail', e.detail);
+					console.groupEnd();
+				}
+				/* eslint-enable no-console */
+			</script>
+
+			<h2>d2l-applied-filters with default label text</h2>
+
+			<d2l-demo-snippet>
+				<d2l-applied-filters for="filter"></d2l-applied-filters>
+			</d2l-demo-snippet>
+
+			<h2>d2l-applied-filters with custom label text</h2>
+
+			<d2l-demo-snippet>
+				<d2l-applied-filters for="filter" label-text="Filters selected:">
+				</d2l-applied-filters>
+			</d2l-demo-snippet>
+
+		</d2l-demo-page>
+
+	</body>
+</html>
