--- demo/d2l-search-facets/d2l-search-facets.html
@@ -1,85 +1,0 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
-		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
-		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-		<script type="module">
-			import '@brightspace-ui/core/components/demo/demo-page.js';
-			import '../../components/d2l-search-facets/d2l-search-facets.js';
-			import '../../components/d2l-search-facets/d2l-search-facets-grouping.js';
-			import '../../components/d2l-search-facets/d2l-search-facets-option.js';
-			import './templated-search-facets-demo.js';
-		</script>
-	</head>
-	<body unresolved>
-
-		<d2l-demo-page page-title="d2l-search-facets">
-
-			<h2>d2l-search-facets</h2>
-
-			<d2l-demo-snippet>
-				<template>
-					<d2l-search-facets>
-						<d2l-search-facets-grouping id="status" value="status" text="My Status" has-more="">
-							<d2l-search-facets-option value="new-courses" text="New Courses Only" count="21"></d2l-search-facets-option>
-							<d2l-search-facets-option value="my-list" text="On My List" count="1"></d2l-search-facets-option>
-						</d2l-search-facets-grouping>
-						<d2l-search-facets-grouping id="format" value="format" text="Format" has-more="">
-							<d2l-search-facets-option value="online" text="Online" count="1"></d2l-search-facets-option>
-							<d2l-search-facets-option value="in-class" text="In-Class" count="2"></d2l-search-facets-option>
-						</d2l-search-facets-grouping>
-						<d2l-search-facets-grouping value="location" text="Location">
-							<d2l-search-facets-option value="outside" text="Outside"></d2l-search-facets-option>
-							<d2l-search-facets-option value="inside" text="Inside"></d2l-search-facets-option>
-						</d2l-search-facets-grouping>
-					</d2l-search-facets>
-					<script>
-						document.getElementById('status').addEventListener('d2l-search-facets-grouping-has-more', (e) => {
-							const options = [
-								{ text: 'In Progress', value: 'in-progress', count: '1' },
-								{ text: 'Completed', value: 'completed', count: '10000' }
-							];
-							addOptions(options, e.target);
-						});
-						document.getElementById('format').addEventListener('d2l-search-facets-grouping-has-more', (e) => {
-							const options = [
-								{ text: 'Interactive', value: 'interactive', count: '12' },
-								{ text: 'Video Only', value: 'video-only', count: '30' }
-							];
-							addOptions(options, e.target);
-						});
-						function addOptions(options, target) {
-							options.forEach((option) => {
-								const element = document.createElement('d2l-search-facets-option');
-								element.setAttribute('text', option.text);
-								element.setAttribute('value', option.value);
-								element.setAttribute('count', option.count);
-								target.appendChild(element);
-							});
-							target.removeAttribute('has-more');
-						}
-					</script>
-				</template>
-			</d2l-demo-snippet>
-
-			<h2>d2l-search-facets (templated)</h2>
-
-			<d2l-demo-snippet>
-				<template>
-					<d2l-labs-templated-search-facets-demo></d2l-labs-templated-search-facets-demo>
-				</template>
-			</d2l-demo-snippet>
-
-		</d2l-demo-page>
-
-		<script>
-			document.addEventListener('d2l-search-facets-change', (e) => {
-				// eslint-disable-next-line no-console
-				console.log(e.detail);
-			});
-		</script>
-
-	</body>
-</html>
