+++ demo/filter-dropdown/filter-dropdown.html
@@ -1,0 +1,102 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script type="module">
+			import '@brightspace-ui/core/components/demo/demo-page.js';
+			import '../../components/filter-dropdown/filter-dropdown.js';
+			import '../../components/filter-dropdown/filter-dropdown-category.js';
+			import '../../components/filter-dropdown/filter-dropdown-option.js';
+			import './templated-filter-dropdown-demo.js';
+		</script>
+		<style>
+			d2l-labs-templated-filter-dropdown-demo, d2l-filter-dropdown { display: inline-block; }
+		</style>
+	</head>
+	<body unresolved>
+
+		<d2l-demo-page page-title="d2l-filter-dropdown">
+
+			<h2>d2l-filter-dropdown with dom-repeat data and event hookups</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-labs-templated-filter-dropdown-demo></d2l-labs-templated-filter-dropdown-demo>
+				</template>
+			</d2l-demo-snippet>
+
+			<h2>Basic d2l-filter-dropdown with hardcoded data and events console logged</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-filter-dropdown id="filter" total-selected-option-count="4">
+						<d2l-filter-dropdown-category key="1" category-text="Category 1">
+							<d2l-filter-dropdown-option selected text="Option 1 - 1 test test test test test test test test test test test test test test test test test test" value="1"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option selected text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option selected text="Option 1 - 4" value="4"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 1 - 5" value="5"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 1 - 6" value="6"></d2l-filter-dropdown-option>
+						</d2l-filter-dropdown-category>
+						<d2l-filter-dropdown-category key="2" category-text="Category 2">
+							<d2l-filter-dropdown-option selected text="Option 2 - 1" value="1"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 2 - 2" value="2"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 2 - 3" value="3"></d2l-filter-dropdown-option>
+						</d2l-filter-dropdown-category>
+						<d2l-filter-dropdown-category key="3" category-text="Category 3">
+							<d2l-filter-dropdown-option text="Option 3 - 1" value="1"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 3 - 2" value="2"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 3 - 3" value="3"></d2l-filter-dropdown-option>
+						</d2l-filter-dropdown-category>
+						<d2l-filter-dropdown-category key="4" category-text="Category 4">
+							<d2l-filter-dropdown-option text="Option 4 - 1" value="1"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 4 - 2" value="2"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 4 - 3" value="3"></d2l-filter-dropdown-option>
+						</d2l-filter-dropdown-category>
+						<d2l-filter-dropdown-category key="5" category-text="Category 5" disable-search>
+							<d2l-filter-dropdown-option text="Option 5 - 1" value="1"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 5 - 2" value="2"></d2l-filter-dropdown-option>
+							<d2l-filter-dropdown-option text="Option 5 - 3" value="3"></d2l-filter-dropdown-option>
+						</d2l-filter-dropdown-category>
+					</d2l-filter-dropdown>
+					<script>
+						const filter = document.querySelector('#filter');
+						filter.addEventListener('d2l-filter-dropdown-close', (e) => {
+							logEvent(e, 'Filter dropdown closed!');
+						});
+						filter.addEventListener('d2l-filter-dropdown-cleared', (e) => {
+							logEvent(e, 'Filters clear button clicked!');
+						});
+						filter.addEventListener('d2l-filter-dropdown-category-selected', (e) => {
+							logEvent(e, 'Filter category tab selected!');
+						});
+						filter.addEventListener('d2l-filter-dropdown-category-searched', (e) => {
+							logEvent(e, 'Filter category searched!');
+						});
+						filter.addEventListener('d2l-filter-dropdown-option-change', (e) => {
+							logEvent(e, 'Filter option clicked!');
+							if (e.detail.selected) {
+								filter.totalSelectedOptionCount++;
+							} else {
+								filter.totalSelectedOptionCount--;
+							}
+						});
+						/* eslint-disable no-console */
+						function logEvent(e, text) {
+							console.group(text);
+							console.log('event', e);
+							if (e.detail) console.log('detail', e.detail);
+							console.groupEnd();
+						}
+						/* eslint-enable no-console */
+					</script>
+				</template>
+			</d2l-demo-snippet>
+
+		</d2l-demo-page>
+
+	</body>
+</html>
