+++ demo/index.html
@@ -14,11 +14,11 @@
 		<d2l-demo-page page-title="demos">
 
 			<ul>
-				<li><a href="d2l-applied-filters/d2l-applied-filters.html">d2l-applied-filters</a></li>
-				<li><a href="d2l-filter-dropdown/d2l-filter-dropdown.html">d2l-filter-dropdown</a></li>
-				<li><a href="d2l-search-facets/d2l-search-facets.html">d2l-search-facets</a></li>
-				<li><a href="d2l-search-results-count/d2l-search-results-count.html">d2l-search-results-count</a></li>
-				<li><a href="d2l-sort-by-dropdown/d2l-sort-by-dropdown.html">d2l-sort-by-dropdown</a></li>
+				<li><a href="applied-filters/applied-filters.html">d2l-applied-filters</a></li>
+				<li><a href="filter-dropdown/filter-dropdown.html">d2l-filter-dropdown</a></li>
+				<li><a href="search-facets/search-facets.html">d2l-search-facets</a></li>
+				<li><a href="search-results-count/search-results-count.html">d2l-search-results-count</a></li>
+				<li><a href="sort-by-dropdown/sort-by-dropdown.html">d2l-sort-by-dropdown</a></li>
 			</ul>
 
 		</d2l-demo-page>
