+++ demo/search-results-count/search-results-count.html
@@ -1,0 +1,56 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script type="module">
+			import '@brightspace-ui/core/components/demo/demo-page.js';
+			import '../../components/search-results-count/search-results-count.js';
+		</script>
+	</head>
+	<body unresolved>
+
+		<d2l-demo-page page-title="d2l-search-results-count">
+
+			<h2>d2l-search-results-count</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-search-results-count total-count="22" query="Financial Planning">
+					</d2l-search-results-count>
+				</template>
+			</d2l-demo-snippet>
+
+			<h2>d2l-search-results-count (no query)</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-search-results-count total-count="22">
+					</d2l-search-results-count>
+				</template>
+			</d2l-demo-snippet>
+
+			<h2>d2l-search-results-count (paged)</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-search-results-count range-start="1" range-end="20" total-count="22" query="Financial Planning">
+					</d2l-search-results-count>
+				</template>
+			</d2l-demo-snippet>
+
+			<h2>d2l-search-results-count (paged, no query)</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-search-results-count range-start="1" range-end="20" total-count="22">
+					</d2l-search-results-count>
+				</template>
+			</d2l-demo-snippet>
+
+		</d2l-demo-page>
+
+	</body>
+</html>
