+++ demo/sort-by-dropdown/sort-by-dropdown.html
@@ -1,0 +1,61 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<link rel="stylesheet" href="/node_modules/@brightspace-ui/core/components/demo/styles.css" type="text/css">
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script type="module">
+			import '@brightspace-ui/core/components/demo/demo-page.js';
+			import '../../components/sort-by-dropdown/sort-by-dropdown-option.js';
+			import '../../components/sort-by-dropdown/sort-by-dropdown.js';
+		</script>
+		<style>
+			d2l-sort-by-dropdown { float: right; }
+		</style>
+	</head>
+	<body unresolved>
+
+		<d2l-demo-page page-title="d2l-sort-by-dropdown">
+
+			<h2>d2l-sort-by-dropdown</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-sort-by-dropdown label="Sort by options" align="end">
+						<d2l-sort-by-dropdown-option value="date" text="Date"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="time" text="Time"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="newest" text="Newest"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="oldest" text="Oldest"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="name" text="Name"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="high-to-low" text="Price: High to Low"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="low-to-high" text="Price: Low to High"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="long-text" text="A sorting option with longer text"></d2l-sort-by-dropdown-option>
+					</d2l-sort-by-dropdown>
+					<div style="clear: both;"></div>
+				</template>
+			</d2l-demo-snippet>
+
+			<h2>d2l-sort-by-dropdown (compact)</h2>
+
+			<d2l-demo-snippet>
+				<template>
+					<d2l-sort-by-dropdown align="end" label="Compact sort by dropdown" compact="">
+						<d2l-sort-by-dropdown-option value="date" text="Date"></d2l-sort-by-dropdown-option>
+						<d2l-sort-by-dropdown-option value="time" text="Time"></d2l-sort-by-dropdown-option>
+					</d2l-sort-by-dropdown>
+					<div style="clear: both;"></div>
+				</template>
+			</d2l-demo-snippet>
+
+		</d2l-demo-page>
+
+		<script>
+			document.addEventListener('d2l-sort-by-dropdown-change', (e) => {
+				// eslint-disable-next-line no-console
+				console.log(e.detail);
+			});
+		</script>
+
+	</body>
+</html>
