--- test/d2l-filter-dropdown/d2l-filter-dropdown.html
@@ -1,167 +1,0 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
-		<title>d2l-filter-dropdown test</title>
-		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-		<script src="/node_modules/wct-browser-legacy/browser.js"></script>
-		<script type="module">
-			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown.js';
-			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown-category.js';
-			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown-option.js';
-		</script>
-	</head>
-	<body>
-		<test-fixture id="basic">
-			<template strip-whitespace>
-				<d2l-filter-dropdown total-selected-option-count="3">
-					<d2l-filter-dropdown-category key="1" category-text="Category 1" selected-option-count="3">
-						<d2l-filter-dropdown-option selected text="Option 1 - 1" value="1"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option selected text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option selected text="Option 1 - 4" value="4"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 1 - 5" value="5"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 1 - 6" value="6"></d2l-filter-dropdown-option>
-					</d2l-filter-dropdown-category>
-					<d2l-filter-dropdown-category key="2" category-text="Category 2" disable-search>
-						<d2l-filter-dropdown-option text="Option 2 - 1" value="1"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 2 - 2" value="2"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 2 - 3" value="3"></d2l-filter-dropdown-option>
-					</d2l-filter-dropdown-category>
-				</d2l-filter-dropdown>
-			</template>
-		</test-fixture>
-
-		<test-fixture id="text-override">
-			<template strip-whitespace>
-				<d2l-filter-dropdown total-selected-option-count="2" header-text="Send To" opener-text="Sending To None" opener-text-single="Sending To One" opener-text-multiple="Sending To Many">
-					<d2l-filter-dropdown-category key="1" category-text="Category 1" selected-option-count="2">
-						<d2l-filter-dropdown-option selected text="Option 1 - 1" value="1"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option selected text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
-					</d2l-filter-dropdown-category>
-				</d2l-filter-dropdown>
-			</template>
-		</test-fixture>
-
-		<test-fixture id="text-variation-disabled">
-			<template strip-whitespace>
-				<d2l-filter-dropdown total-selected-option-count="2" disable-opener-text-variation>
-					<d2l-filter-dropdown-category key="1" category-text="Category 1" selected-option-count="2">
-						<d2l-filter-dropdown-option selected text="Option 1 - 1" value="1"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
-						<d2l-filter-dropdown-option selected text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
-					</d2l-filter-dropdown-category>
-				</d2l-filter-dropdown>
-			</template>
-		</test-fixture>
-
-		<script type="module">
-			import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
-
-			suite('d2l-filter-dropdown', () => {
-				let filter;
-
-				setup((done) => {
-					filter = fixture('basic');
-					afterNextRender(filter, done);
-				});
-				test('instantiating the element works', () => {
-					assert.equal('d2l-filter-dropdown', filter.tagName.toLowerCase());
-				});
-				test('attributes are set correctly', () => {
-					assert.equal(filter.totalSelectedOptionCount, 3);
-				});
-				test('clear button is hidden if total selected count is 0, displayed otherwise', (done) => {
-					const clearButton = filter.shadowRoot.querySelector('d2l-button-subtle');
-					assert.isFalse(clearButton.hidden);
-
-					filter.totalSelectedOptionCount = 0;
-
-					requestAnimationFrame(() => {
-						assert.isTrue(clearButton.hidden);
-						done();
-					});
-				});
-				test('pressing the clear button triggers the d2l-filter-dropdown-cleared event', (done) => {
-					const clearButton = filter.shadowRoot.querySelector('d2l-button-subtle');
-					filter.addEventListener('d2l-filter-dropdown-cleared', () => {
-						done();
-					});
-
-					clearButton.click();
-				});
-				test('closing the filter triggers the d2l-filter-dropdown-close event', (done) => {
-					const dropdown = filter.shadowRoot.querySelector('d2l-dropdown-tabs');
-					filter.addEventListener('d2l-filter-dropdown-close', () => {
-						done();
-					});
-
-					dropdown.toggleOpen();
-
-					requestAnimationFrame(() => {
-						requestAnimationFrame(() => {
-							dropdown.toggleOpen();
-						});
-					});
-				});
-				test('the opener text is set accordingly', () => {
-					const openerButton = filter.shadowRoot.querySelector('d2l-dropdown-button-subtle');
-					assert.equal(openerButton.text, 'Filter: 3 Filters');
-
-					filter.totalSelectedOptionCount = 1;
-					assert.equal(openerButton.text, 'Filter: 1 Filter');
-
-					filter.totalSelectedOptionCount = 0;
-					assert.equal(openerButton.text, 'Filter');
-				});
-				test('filter categories display correctly', (done) => {
-					function getTabs() {
-						const tabs = filter.shadowRoot.querySelector('d2l-tabs').shadowRoot.querySelectorAll('d2l-tab-internal');
-						if (tabs.length === 2) {
-							assert.equal(tabs[0].text, 'Category 1 (3)');
-							assert.equal(tabs[1].text, 'Category 2');
-							done();
-						} else {
-							setTimeout(getTabs, 25);
-						}
-					}
-
-					getTabs();
-				});
-				test('focussing the filter sets the focus on the opener button', () => {
-					assert.equal(filter.shadowRoot.activeElement, null);
-					filter.focus();
-					assert.equal(filter.shadowRoot.activeElement.tagName.toLowerCase(), 'd2l-dropdown-button-subtle');
-				});
-				test('the text values can be overridden', () => {
-					filter = fixture('text-override');
-					const header = filter.shadowRoot.querySelector('.d2l-filter-dropdown-content-header span');
-					assert.equal(header.innerText, 'Send To');
-
-					const openerButton = filter.shadowRoot.querySelector('d2l-dropdown-button-subtle');
-					assert.equal(openerButton.text, 'Sending To Many');
-
-					filter.totalSelectedOptionCount = 1;
-					assert.equal(openerButton.text, 'Sending To One');
-
-					filter.totalSelectedOptionCount = 0;
-					assert.equal(openerButton.text, 'Sending To None');
-				});
-
-				test('the opener text variations can be disabled', () => {
-					filter = fixture('text-variation-disabled');
-					const openerButton = filter.shadowRoot.querySelector('d2l-dropdown-button-subtle');
-					assert.equal(openerButton.text, 'Filter');
-
-					filter.totalSelectedOptionCount = 1;
-					assert.equal(openerButton.text, 'Filter');
-
-					filter.totalSelectedOptionCount = 0;
-					assert.equal(openerButton.text, 'Filter');
-				});
-			});
-		</script>
-	</body>
-</html>
