+++ test/filter-dropdown/filter-dropdown-option.html
@@ -1,0 +1,99 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<title>d2l-filter-dropdown-option test</title>
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script src="/node_modules/wct-browser-legacy/browser.js"></script>
+		<script type="module">
+			import '@brightspace-ui/core/components/menu/menu.js';
+			import '../../components/filter-dropdown/filter-dropdown-option.js';
+		</script>
+	</head>
+	<body>
+		<test-fixture id="basic">
+			<template strip-whitespace>
+				<d2l-menu>
+					<d2l-filter-dropdown-option text="Option 1 - 1" value="1"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option text="Option 1 - 2" value="2"></d2l-filter-dropdown-option>
+					<d2l-filter-dropdown-option selected text="Option 1 - 3" value="3"></d2l-filter-dropdown-option>
+				</d2l-menu>
+			</span>
+			</template>
+		</test-fixture>
+
+		<script type="module">
+			import '@polymer/iron-test-helpers/mock-interactions.js';
+			import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
+
+			suite('d2l-filter-dropdown-option', () => {
+				let container;
+				let options = [];
+
+				setup((done) => {
+					container = fixture('basic');
+					options = container.querySelectorAll('d2l-filter-dropdown-option');
+					afterNextRender(container, done);
+				});
+				test('instantiating the element works', () => {
+					assert.equal('d2l-filter-dropdown-option', options[0].tagName.toLowerCase());
+				});
+				test('role is set properly', () => {
+					assert.equal(options[0].getAttribute('role'), 'menuitemcheckbox');
+
+				});
+				test('fires d2l-menu-item-change event when item is clicked', (done) => {
+					options[0].addEventListener('d2l-menu-item-change', (e) => {
+						assert.equal(e.detail.value, '1');
+						assert.equal(e.detail.selected, true);
+						done();
+					});
+					options[0].click();
+				});
+				test('fires d2l-menu-item-change event when enter key pressed on item', (done) => {
+					options[0].addEventListener('d2l-menu-item-change', (e) => {
+						assert.equal(e.detail.value, '1');
+						assert.equal(e.detail.selected, true);
+						done();
+					});
+					MockInteractions.keyDownOn(options[0], 13);
+				});
+				test('fires d2l-menu-item-change event when space key pressed on item', (done) => {
+					options[0].addEventListener('d2l-menu-item-change', (e) => {
+						assert.equal(e.detail.value, '1');
+						assert.equal(e.detail.selected, true);
+						done();
+					});
+					MockInteractions.keyDownOn(options[0], 32);
+				});
+				test('should toggle state on selection', (done) => {
+					let expected = true;
+					options[0].addEventListener('d2l-menu-item-change', (e) => {
+						assert.equal(options[0].selected, expected);
+						assert.equal(e.detail.selected, expected);
+						assert.equal(e.detail.value, '1');
+						options[0].updateComplete.then(() => {
+							assert.equal(window.getComputedStyle(options[0].shadowRoot.querySelector('d2l-icon')).getPropertyValue('visibility'), expected ? 'visible' : 'hidden');
+							if (expected) {
+								expected = false;
+								options[0].click();
+							} else {
+								done();
+							}
+						});
+					});
+					options[0].click();
+				});
+				test('does not affect other checkboxes in the menu when selected', (done) => {
+					options[0].addEventListener('d2l-menu-item-change', () => {
+						assert.equal(options[0].selected, true);
+						assert.notEqual(options[1].selected, true);
+						done();
+					});
+					options[0].click();
+				});
+			});
+		</script>
+	</body>
+</html>
