+++ test/index.html
@@ -10,15 +10,15 @@
 	<body>
 		<script>
 			WCT.loadSuites([
-				'd2l-search-facets/d2l-search-facets.html',
-				'd2l-search-facets/d2l-search-facets-option.html',
-				'd2l-search-facets/d2l-search-facets-grouping.html',
-				'd2l-search-results-count/d2l-search-results-count.html',
-				'd2l-sort-by-dropdown/d2l-sort-by-dropdown.html',
-				'd2l-filter-dropdown/d2l-filter-dropdown.html',
-				'd2l-filter-dropdown/d2l-filter-dropdown-category.html',
-				'd2l-filter-dropdown/d2l-filter-dropdown-option.html',
-				'd2l-applied-filters/d2l-applied-filters.html'
+				'search-facets/search-facets.html',
+				'search-facets/search-facets-option.html',
+				'search-facets/search-facets-grouping.html',
+				'search-results-count/search-results-count.html',
+				'sort-by-dropdown/sort-by-dropdown.html',
+				'filter-dropdown/filter-dropdown.html',
+				'filter-dropdown/filter-dropdown-category.html',
+				'filter-dropdown/filter-dropdown-option.html',
+				'applied-filters/applied-filters.html'
 			]);
 		</script>
 	</body>
