+++ test/search-facets/search-facets.html
@@ -1,0 +1,85 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<title>d2l-search-facets test</title>
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script src="/node_modules/wct-browser-legacy/browser.js"></script>
+		<script type="module">
+			import '../../components/search-facets/search-facets.js';
+			import '../../components/search-facets/search-facets-grouping.js';
+			import '../../components/search-facets/search-facets-option.js';
+		</script>
+	</head>
+	<body>
+		<test-fixture id="basic">
+			<template>
+				<d2l-search-facets>
+					<d2l-search-facets-grouping id="grouping1" value="test-grouping1" text="Grouping 1">
+						<d2l-search-facets-option
+							id="grouping1-option1"
+							value="grouping1-option1"
+							text="Grouping 1 Option 1"
+							count="1"
+						></d2l-search-facets-option>
+						<d2l-search-facets-option
+							id="grouping1-option2"
+							value="grouping1-option2"
+							text="Grouping 1 Option 2"
+							count="2"
+						></d2l-search-facets-option>
+					</d2l-search-facets-grouping>
+				</d2l-search-facets>
+			</template>
+		</test-fixture>
+		<script type="module">
+			import './test-helper-functions.js';
+			import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
+
+			describe('d2l-search-facets', () => {
+				describe('basic', () => {
+					let basicFixture;
+					let grouping1;
+					let grouping1Option1;
+					let grouping1Option2;
+
+					beforeEach((done) => {
+						basicFixture = fixture('basic');
+						afterNextRender(basicFixture, () => {
+							grouping1 = document.getElementById('grouping1');
+							grouping1Option1 = document.getElementById('grouping1-option1');
+							grouping1Option2 = document.getElementById('grouping1-option2');
+							done();
+						});
+					});
+
+					it('should emit an event when an option is selected', (done) => {
+						let selectCount = 0;
+						basicFixture.addEventListener('d2l-search-facets-change', (e) => {
+							selectCount += 1;
+
+							if (selectCount === 1) {
+								expect(e.detail).to.deep.equal({
+									checked: true,
+									option: grouping1Option1.value,
+									grouping: grouping1.value,
+								});
+							} else {
+								expect(e.detail).to.deep.equal({
+									checked: true,
+									grouping: grouping1.value,
+									option: grouping1Option2.value
+								});
+								done();
+							}
+						});
+
+						selectOption(grouping1Option1);
+						selectOption(grouping1Option2);
+					});
+				});
+			});
+</script>
+	</body>
+</html>
