+++ test/search-facets/test-helper-functions.js
