+++ test/search-results-count/search-results-count.html
@@ -1,0 +1,100 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<title>d2l-search-results-count test</title>
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script src="/node_modules/wct-browser-legacy/browser.js"></script>
+		<script type="module">
+			import '../../components/search-results-count/search-results-count.js';
+		</script>
+	</head>
+	<body>
+		<test-fixture id="basic">
+			<template>
+				<d2l-search-results-count total-count="22" query="Financial Planning"></d2l-search-results-count>
+			</template>
+		</test-fixture>
+		<test-fixture id="basic-no-query">
+			<template>
+				<d2l-search-results-count total-count="22"></d2l-search-results-count>
+			</template>
+		</test-fixture>
+		<test-fixture id="paged">
+			<template>
+				<d2l-search-results-count range-start="1" range-end="20" total-count="22" query="Financial Planning"></d2l-search-results-count>
+			</template>
+		</test-fixture>
+		<test-fixture id="paged-no-query">
+			<template>
+				<d2l-search-results-count range-start="1" range-end="20" total-count="22"></d2l-search-results-count>
+			</template>
+		</test-fixture>
+		<script type="module">
+			import '@polymer/iron-test-helpers/mock-interactions.js';
+			import '@polymer/iron-test-helpers/test-helpers.js';
+			import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
+			describe('d2l-search-results-count', () => {
+				describe('basic', () => {
+					let basicFixture;
+
+					beforeEach((done) => {
+						basicFixture = fixture('basic');
+						afterNextRender(basicFixture, done);
+					});
+
+					it('text should match', () => {
+						const shadowRoot = basicFixture.shadowRoot;
+						const text = shadowRoot.textContent.replace(/[\n\r]+|[\s]{2,}/g, ' ').trim();
+						expect(text).to.equal('22 results for "Financial Planning"');
+					});
+				});
+
+				describe('basic no query', () => {
+					let basicFixture;
+
+					beforeEach((done) => {
+						basicFixture = fixture('basic-no-query');
+						afterNextRender(basicFixture, done);
+					});
+
+					it('text should match', () => {
+						const shadowRoot = basicFixture.shadowRoot;
+						const text = shadowRoot.textContent.replace(/[\n\r]+|[\s]{2,}/g, ' ').trim();
+						expect(text).to.equal('22 results');
+					});
+				});
+				describe('paged', () => {
+					let basicFixture;
+
+					beforeEach((done) => {
+						basicFixture = fixture('paged');
+						afterNextRender(basicFixture, done);
+					});
+
+					it('text should match', () => {
+						const shadowRoot = basicFixture.shadowRoot;
+						const text = shadowRoot.textContent.replace(/[\n\r]+|[\s]{2,}/g, ' ').trim();
+						expect(text).to.equal('1 - 20 of 22 results for "Financial Planning"');
+					});
+				});
+
+				describe('paged no query', () => {
+					let basicFixture;
+
+					beforeEach((done) => {
+						basicFixture = fixture('paged-no-query');
+						afterNextRender(basicFixture, done);
+					});
+
+					it('text should match', () => {
+						const shadowRoot = basicFixture.shadowRoot;
+						const text = shadowRoot.textContent.replace(/[\n\r]+|[\s]{2,}/g, ' ').trim();
+						expect(text).to.equal('1 - 20 of 22 results');
+					});
+				});
+			});
+		</script>
+	</body>
+</html>
