+++ test/sort-by-dropdown/sort-by-dropdown.html
@@ -1,0 +1,57 @@
+<!doctype html>
+<html lang="en">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
+		<title>d2l-sort-by-dropdown test</title>
+		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+		<script src="/node_modules/wct-browser-legacy/browser.js"></script>
+		<script type="module">
+			import '../../components/sort-by-dropdown/sort-by-dropdown-option.js';
+			import '../../components/sort-by-dropdown/sort-by-dropdown.js';
+		</script>
+	</head>
+	<body>
+		<test-fixture id="basic">
+			<template>
+				<d2l-sort-by-dropdown label="Sort by test">
+					<d2l-sort-by-dropdown-option id="option1" value="option1" text="Option 1"></d2l-sort-by-dropdown-option>
+					<d2l-sort-by-dropdown-option id="option2" value="option2" text="Option 2"></d2l-sort-by-dropdown-option>
+				</d2l-sort-by-dropdown>
+			</template>
+		</test-fixture>
+		<script type="module">
+			import '@polymer/iron-test-helpers/mock-interactions.js';
+			import '@polymer/iron-test-helpers/test-helpers.js';
+			import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
+			describe('d2l-sort-by-dropdown', () => {
+				describe('basic', () => {
+					let basicFixture;
+					let option1;
+					let option2;
+
+					beforeEach((done) => {
+						basicFixture = fixture('basic');
+						afterNextRender(basicFixture, () => {
+							option1 = basicFixture.querySelector('#option1');
+							option2 = basicFixture.querySelector('#option2');
+							done();
+						});
+					});
+
+					it('should initially select the first option by default', () => {
+						expect(basicFixture.value).to.equal('option1');
+					});
+
+					it('should change to the selected option', () => {
+						MockInteractions.tap(option2);
+						expect(basicFixture.getAttribute('value')).to.equal(option2.value);
+
+						MockInteractions.tap(option1);
+						expect(basicFixture.getAttribute('value')).to.equal(option1.value);
+					});
+				});
+			});
+		</script>
+	</body>
+</html>
